pyproject.toml
[project]
name = "server"
version = "0.0.1"
description = ""
authors = [
    {name = "songbinghui", email = "soongbinghui@gmail.com"},
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0.post1",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.2",
    "pytest>=7.4.3",
    "openai>=1.3.8",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
dev = "python src/server/main.py"
test = "pytest"

[tool.ruff]
exclude = [
]
